:art: Moving logger file to utils.
